ArcGIS service/layer support Flesh out service detail edit page
Modify geonetwork - use 2.6.x branch from github, must set all records to public (csw server configuration)
This is to prevent geonetwork from crashing after importing several thousand records
After about ~1600 rapid insert transactions to geonetwork 2.6.4, tomcat hangs:
“Too many open files” and/or
All database connections locked up
http://trac.osgeo.org/geonetwork/ticket/1098
Build geonetwork 2.6.x from source, includes patches for this problem, but:
this breaks the code that tries to make record publicly visible after insert
solution: make all records visible by default from geonetwork admin console (todo: set this in configuration file somehow?), and comment out that part of the code
Also tried out geonetwork 2.10, required many changes to code, got it to work but will stick with 2.6.x for now until further testing can be done

+++ geonode/contrib/services/tasks.py
@@ -5,29 +5,19 @@ from geonode.contrib.services.models import WebServiceHarvestLayersJob, WebServi
 from geonode.contrib.services.views import update_layers, register_service_by_type, _register_indexed_layers
 from geonode.maps import autocomplete_light_registry
 
-
-@task
-def import_indexed_wms_service_layers(service, wms=None, owner=None):
-    _register_indexed_layers(service, wms=wms, owner=owner)
-
-
-@task
-def import_indexed_service(url, type, username=None, password=None, owner=None):
-    register_service_by_type(url, type, username=username, password=password, owner=owner)
-
 @periodic_task(run_every=crontab(minute=settings.SERVICE_UPDATE_INTERVAL))
 def harvest_service_layers():
-    harvestJobs = WebServiceHarvestLayersJob.objects.all()
-    for job in harvestJobs.exclude(status="process"):
-        try:
-            job.status = "process"
-            job.save()
-            update_layers(job.service)
-            job.delete()
-        except Exception, e:
-            print e
-            job.status = 'failed'
-            job.save()
+    if WebServiceHarvestLayersJob.objects.filter(status="process").count() == 0:
+        for job in WebServiceHarvestLayersJob.objects.exclude(status="process").exclude(status="done"):
+            try:
+                job.status = "process"
+                job.save()
+                update_layers(job.service)
+                job.delete()
+            except Exception, e:
+                print e
+                job.status = 'failed'
+                job.save()
 
 @periodic_task(run_every=crontab(minute=settings.SERVICE_UPDATE_INTERVAL))
 def import_service():
